chore: Update application version to v3.7
This commit updates the application version string in index.html from "v3.6" to "v3.7". This is a minor version bump, likely for tracking release changes. The deployment workflow file `.github/workflows/static.yml` was also removed as it seems to be a remnant or no longer necessary for the current project setup.